recrutadorai_webapp_final/backend/requirements.txt
fastapi
uvicorn
python-docx
PyPDF2
scikit-learn
sentence-transformers
python-multipart
